Handle tm_utcoff correctly on Windows and in to_timespec.
Even during daylight savings time, Tm.tm_utcoff would be populated with
the non-daylight savings offset. Making the Windows-specific code call the
functions that correspond more closely to the Unix-specific functions
resolves that.

to_timespec ignored tm_utcoff except for checking whether it was 0, which
caused this to be benign except in the case of a time zone where standard
is UTC+0 and DST is UTC+-1. It now takes tm_utcoff into account.